Update incorporating changes in NPL for review comments - part13
- Improvement in Logging.
- Added capability to update Pods in clusterrole.
