+++ pkg/agent/nodeportlocal/k8s/controller.go
@@ -17,6 +17,7 @@
 package k8s
 
 import (
+	"encoding/json"
 	"fmt"
 	"time"
 
@@ -55,10 +56,8 @@ func NewNPLController(kubeClient clientset.Interface, pt *portcache.PortTable) *
 	return &nplCtrl
 }
 
-// Run starts to watching and process Pod updates for the Node where Antrea Agent is running.
-// It starts a fixed numbers of queues and one worker per queue to process the objects.
-// Each object is mapped to one queue based on the hash of the object key. This ensures that
-// update for one object always gets processes by the same worker.
+// Run starts to watch and process Pod updates for the Node where Antrea Agent is running.
+// It starts a queue and a fixed number of workers to process the objects from the queue.
 func (c *Controller) Run(stopCh <-chan struct{}) {
 	defer c.queue.ShutDown()
 	for i := 0; i < numWorkers; i++ {
@@ -159,7 +158,105 @@ func (c *Controller) processNextWorkItem() bool {
 		c.queue.Forget(key)
 	} else {
 		c.queue.AddRateLimited(key)
-		klog.Errorf("Error syncing pod %s, requeuing. Error: %v", key, err)
+		klog.Errorf("Error syncing Pod %s, requeuing. Error: %v", key, err)
 	}
 	return true
+}
+
+func (c *Controller) addRuleForPod(pod *corev1.Pod) error {
+	podIP, nodeIP := pod.Status.PodIP, pod.Status.HostIP
+	if podIP == "" || nodeIP == "" {
+		return nil
+	}
+	podContainers := pod.Spec.Containers
+
+	for _, container := range podContainers {
+		for _, cport := range container.Ports {
+			port := int(cport.ContainerPort)
+			if c.portTable.RuleExists(podIP, port) {
+				continue
+			}
+			nodePort, err := c.portTable.AddRule(podIP, port)
+			if err != nil {
+				return err
+			}
+			assignPodAnnotation(pod, nodeIP, port, nodePort)
+		}
+	}
+	return nil
+}
+
+// HandleDeletePod Removes rules from port table and
+// rules programmed in the system based on implementation type (e.g. IPTABLES)
+func (c *Controller) HandleDeletePod(key string) error {
+	klog.Infof("Got delete event for Pod: %s", key)
+	podIP, found := c.PodToIP[key]
+	if !found {
+		klog.Infof("IP address not found for Pod: %s", key)
+		return nil
+	}
+	data := c.portTable.GetDataForPodIP(podIP)
+	for _, d := range data {
+		err := c.portTable.DeleteRule(d.PodIP, int(d.PodPort))
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// HandleAddUpdatePod handles Pod Add, Update events and updates annotation if required
+func (c *Controller) HandleAddUpdatePod(key string, obj interface{}) error {
+	newPod := obj.(*corev1.Pod).DeepCopy()
+	klog.Infof("Got add/update event for pod: %s", key)
+
+	podIP := newPod.Status.PodIP
+	if podIP == "" {
+		klog.Infof("IP address not found for pod: %s/%s", newPod.Namespace, newPod.Name)
+		return nil
+	}
+	c.PodToIP[key] = podIP
+
+	var err error
+	var updatePodAnnotation bool
+	newPodPorts := make(map[int]struct{})
+	newPodContainers := newPod.Spec.Containers
+	for _, container := range newPodContainers {
+		for _, cport := range container.Ports {
+			port := int(cport.ContainerPort)
+			newPodPorts[port] = struct{}{}
+			if !c.portTable.RuleExists(podIP, int(cport.ContainerPort)) {
+				err = c.addRuleForPod(newPod)
+				if err != nil {
+					return fmt.Errorf("Failed to add rule for Pod %s/%s: %s", newPod.Namespace, newPod.Name, err.Error())
+				}
+				updatePodAnnotation = true
+			}
+		}
+	}
+
+	var annotations []NPLAnnotation
+	podAnnotation := newPod.GetAnnotations()
+	entries := c.portTable.GetDataForPodIP(podIP)
+	if podAnnotation != nil {
+		if err := json.Unmarshal([]byte(podAnnotation[NPLAnnotationStr]), &annotations); err != nil {
+			klog.Warningf("Unable to unmarshal NodePorLocal annotation")
+			return nil
+		}
+		for _, data := range entries {
+			if _, exists := newPodPorts[data.PodPort]; !exists {
+				removeFromPodAnnotation(newPod, data.PodPort)
+				err := c.portTable.DeleteRule(podIP, int(data.PodPort))
+				if err != nil {
+					return fmt.Errorf("Failed to delete rule for Pod IP %s, Pod Port %d: %s", podIP, data.PodPort, err.Error())
+				}
+				updatePodAnnotation = true
+			}
+		}
+	}
+
+	if updatePodAnnotation {
+		return c.updatePodAnnotation(newPod)
+	}
+	return nil
 }
